Added barrelling for track progress
