+++ src/components/MediaControlsBar/TrackProgress/index.tsx
@@ -1,40 +1,1 @@
-import React, { ReactNode } from "react";
-import { TIME_FORMAT } from "../../../constants";
-import CurrentTime from "./CurrentTime";
-import Duration from "./Duration";
-import ProgressBar from "./ProgressBar";
-import { I18nAriaLabels } from "../../../types";
-
-interface TrackProgressProps {
-  audio: HTMLAudioElement | null;
-  defaultCurrentTime?: ReactNode;
-  defaultDuration?: ReactNode;
-  timeFormat?: TIME_FORMAT;
-  i18nAriaLabels?: I18nAriaLabels;
-}
-
-function TrackProgress({ audio, timeFormat, defaultCurrentTime, defaultDuration, i18nAriaLabels }: TrackProgressProps) {
-  const progressRef = React.useRef<HTMLDivElement>(null);
-  if (!timeFormat) {
-    return null;
-  }
-  return (
-    <div className="w-full flex flex-row items-center justify-evenly gap-2">
-      <CurrentTime
-        audio={audio}
-        className="min-w-[2.5rem] text-xs text-gray-700 dark:text-gray-350 text-right"
-        defaultCurrentTime={defaultCurrentTime}
-        timeFormat={timeFormat}
-      />
-      <ProgressBar audio={audio} i18nProgressBar={i18nAriaLabels?.progressControl} ref={progressRef} />
-      <Duration
-        audio={audio}
-        className="min-w-[2.5rem] text-xs text-gray-700 dark:text-gray-350 text-left"
-        defaultDuration={defaultDuration}
-        timeFormat={timeFormat}
-      />
-    </div>
-  );
-}
-
-export default TrackProgress;
+export { default } from "./TrackProgress";
